requirements.txt
fastapi[all]~=0.109.2
uvicorn[standard]
pydantic~=2.6.1
SQLAlchemy~=2.0.25
python-dotenv~=1.0.1
bcrypt~=4.1.2
pydantic_settings~=2.1.0
